Move `PlanetAttributes` to `libOPHD`
