+++ libOPHD/libOPHD.vcxproj
@@ -164,6 +164,7 @@
   <ItemGroup>
     <ClCompile Include="DirectionOffset.cpp" />
     <ClCompile Include="MeanSolarDistance.cpp" />
+    <ClCompile Include="PlanetAttributes.cpp" />
     <ClCompile Include="ProductCatalogue.cpp" />
     <ClCompile Include="StorableResources.cpp" />
     <ClCompile Include="XmlSerializer.cpp" />
@@ -190,6 +191,7 @@
     <ClInclude Include="EnumStructureID.h" />
     <ClInclude Include="EnumTerrainType.h" />
     <ClInclude Include="MeanSolarDistance.h" />
+    <ClInclude Include="PlanetAttributes.h" />
     <ClInclude Include="ProductCatalogue.h" />
     <ClInclude Include="ProductionCost.h" />
     <ClInclude Include="RandomNumberGenerator.h" />
